requirements.txt
aiogram
discord.py
